- .cw level loading - Custom block definitions and permissions

+++ org/opencraft/server/game/GameModeAdapter.java
@@ -103,7 +103,7 @@ public abstract class GameModeAdapter<P extends Player> implements GameMode<P> {
 
   // Default implementation
   public void setBlock(Player player, Level level, int x, int y, int z, int mode, int type) {
-    level.setBlock(x, y, z, (byte) (mode == 1 ? type : 0));
+    level.setBlock(x, y, z, (mode == 1 ? type : 0));
   }
 
   // Default implementation
